Ignore small time changes in alarm manager.
It turns out switching to the new kernel alarm
reporting causes lots and lots of spurious flags
that the clock has changed.  The alarm manager
would blindly trust these, thinking the world has
changed on it and recomputing everything and reporting
this to everyone else.  This was expensive.

We now verify that the time has changed sufficiently
that it is worth caring about.  This is basically the
same algorithm that battery stats was using to avoid
recording small clock changes, so we are really just
pushing this down into the alarm manager and can now
remove that from battery stats.

Also since we are getting these so much, make use of
the other information in about the wakeup that tells us
if an alarm went off to avoid doing anything if it didn't.

Change-Id: I6f4f4226db6eb2b38ca73860786e7cf7c9136cc3