Camera2: Fix session close callback
If a session is directly replaced, it no longer will receive state
change callbacks from the device. That means it'll never hear about
abort or idle transitions.

Fix this to signal onClose earlier, after sequences complete. This is
reliable to wait on.

Remove the unconfigure drainer, as we have nothing to do after
null-stream configuration completes, now, so no point in triggering
anything for it.

Bug: 19666551
Change-Id: Ied6fb40126b79886cdbcb27d40a76a1646273dd2