Postpone AnimatedVectorDrawable animator inflation until applyTheme()
This CL works around Animator's lack of support for applying a theme
after inflation by postponing Animator inflation until a theme is
available, either in inflate() or applyTheme().

Includes a workaround for AVDs that don't reference any theme attrs
in their animators and this don't require a call to applyTheme().

Partial implementation of removing non-constant data from the AVD's
constant state. Moves the mutable AnimatorSet to a local variable and
treats the Animators as constant data.

We'll follow up with real support for applyTheme() in Animator or
AnimatorSet, at which point we can remove this workaround.

Bug: 20817800
Change-Id: I555c53c955219990ee376bee568bcc038532f9ed