TIF: Generate hardware input id in language-neutral fashion
For certain RtoL languages, digit used as a part of input id
(like HW0) is also localized, hence language change results in
the duplicated inputs for an identical hardware. This CL
changes the way the id is built to language-neutral fashion,
and keeps the digit used in the id in Arabic.

Bug: 21523750
Change-Id: I8f95e0209eeee16ceb543e574ef9ed572e4d031f