drawBitmap() support for extractSubset()
In SkiaCanvasProxy, translate drawBitmap() calls on bitmaps which
are the result of extractSubset() into drawBitmapRect() calls on the
original PixelRef.

If HWUI isn't going to support subsetting, we'll need to do this,
probably at additional entry points, to interoperate with native Skia
bitmaps. This change by itself fixes the extractbitmap GM.

R=djsollen@google.com
BUG=skia:3606

Change-Id: I5744da87d8f16dc1a252f606dfaa4016bd01e66e