Add support for efficient copy within a document provider.
Currently copying files even within the same document provider causes
reading and writing all of the bytes, which is very inefficient for all
network based solutions, eg. Drive where copying even huge files can be
instant.

This CL adds an optional copyDocument method to the DocumentProvider as
well as DocumentContract so providers can implement an efficient way of
copying things. If not implemented, we would fallback to copying by
streaming.

BUG=None

Change-Id: I8f2b2c0c834717f07d42f8247cd1fc025e82239a