Make READ_EXTERNAL_STORAGE dangerous
We have a new storage permission group that has read and
write external storage. However, read external storage is
(not a regression) a normal permission while write is a
dangerous one. This leads to cases where the user disables
the Storage permission and apps still read form it. This
change makes read external storage a dangerous permission.

bug:21949045

Change-Id: I7e28f629dda6e9c1f70cb20a3d5bea74fb109890