Suppress pausing activities when screen off due to proximity.
Pausing and resuming the in-call activity may have the side-effect
of releasing and reacquiring the proximity sensor wake lock which can
cause the system to enter a loop.

Although in general it might be a good idea to pause activities
whenever you can't see them, it makes more sense to treat the proximity
sensor as temporarily suppressing rather than fully pausing activities.

Bug: 21932251
Change-Id: Ie072cbe5984bec08f2c60c64ed8cc144534982db