Move tuner enable point to QS
Long clicking on settings gear in QS causes it to accelarate and
spin, when released it will open and enable the tuner.  On first
opening there will be a warning dialog.

Also add way to disable tuner from tuner screen. This way it can
be removed from its former home in Developer Settings.  All tuner
settings are reset when disabled to match developer settings
behavior.

Bug: 22462605
Change-Id: Ie700a6a15e6c3caccf7cd1885da73328e6fac6ab